dificult

+++ db/migrate/20211006141922_create_bookmarks.rb
@@ -1,8 +1,8 @@
 class CreateBookmarks < ActiveRecord::Migration[5.2]
   def change
     create_table :bookmarks do |t|
-      t.references :user, type: :integer, foreign_key: true
-      t.references :post, type: :integer, foreign_key: true
+      t.references :user, type: :bigint, foreign_key: true
+      t.references :post, type: :bigint, foreign_key: true
 
       t.timestamps
     end
